Added a PointType to CarbonDataPoint

+++ Greenfoot/Core Data Models/CarbonDataPoint+CoreDataClass.swift
@@ -17,6 +17,7 @@ public class CarbonDataPoint: NSManagedObject, CoreDataRecord {
     required convenience init?(inContext context: NSManagedObjectContext,
                               source: CarbonSource,
                               unit: CarbonUnit,
+                              type: PointType,
                               month: NSDate,
                               value: Double) {
         self.init(context: context)
@@ -25,17 +26,28 @@ public class CarbonDataPoint: NSManagedObject, CoreDataRecord {
         self.month = month
         self.lastUpdated = NSDate()
         self.rawValue = value
+        self.pointType = type
         
-        guard let carbonConversion = self.unit.carbonConversion else {
-            print("Couldn't create CarbonDataPoint because Unit has no conversions")
-            context.delete(self)
-            return nil
+        if type == .direct {
+            guard let carbonConversion = self.unit.carbonConversion else {
+                print("Couldn't create CarbonDataPoint because Unit has no conversions")
+                context.delete(self)
+                return nil
+            }
+            
+            self.carbonValue = carbonConversion.convert(value)
         }
-        
-        self.carbonValue = carbonConversion.convert(value)
     }
     
     func reference(atLevel level: CarbonReference.Level) -> CarbonReference? {
         return self.references?.first(where: {($0 as? CarbonReference)?.level == level}) as? CarbonReference
+    }
+}
+
+extension CarbonDataPoint {
+    @objc
+    public enum PointType: Int16 {
+        case direct = 0,
+        derived = 1
     }
 }
